zack comments
