+++ cli/auth.go
@@ -7,6 +7,7 @@ import (
 	"io"
 	"net/http"
 	"net/url"
+	"os"
 	"os/exec"
 	"runtime"
 	"strconv"
@@ -17,7 +18,11 @@ import (
 	"github.com/golang-jwt/jwt/v4"
 	"github.com/pkg/errors"
 	"github.com/urfave/cli/v2"
+	datapb "go.viam.com/api/app/data/v1"
+	apppb "go.viam.com/api/app/v1"
+
 	"go.viam.com/utils"
+	"go.viam.com/utils/rpc"
 )
 
 type authFlow struct {
@@ -145,8 +150,8 @@ func PrintAccessTokenAction(c *cli.Context) error {
 		return err
 	}
 
-	if client.conf.Auth == nil || client.conf.Auth.isExpired() {
-		return errors.New("not logged in. run \"login\" command")
+	if err := client.ensureLoggedIn(); err != nil {
+		return err
 	}
 
 	fmt.Fprintln(c.App.Writer, client.conf.Auth.AccessToken)
@@ -184,6 +189,98 @@ func WhoAmIAction(c *cli.Context) error {
 	}
 	fmt.Fprintf(c.App.Writer, "%s\n", auth.User.Email)
 	return nil
+}
+
+func (c *appClient) ensureLoggedIn() error {
+	if c.client != nil {
+		return nil
+	}
+
+	if c.conf.Auth == nil {
+		return errors.New("not logged in: run the following command to login:\n\tviam login")
+	}
+
+	if c.conf.Auth.isExpired() {
+		if !c.conf.Auth.canRefresh() {
+			utils.UncheckedError(c.logout())
+			return errors.New("token expired and cannot refresh")
+		}
+
+		// expired.
+		newToken, err := c.authFlow.refreshToken(c.c.Context, c.conf.Auth)
+		if err != nil {
+			utils.UncheckedError(c.logout()) // clear cache if failed to refresh
+			return errors.Wrapf(err, "error while refreshing token")
+		}
+
+		// write token to config.
+		c.conf.Auth = newToken
+		if err := storeConfigToCache(c.conf); err != nil {
+			return err
+		}
+	}
+
+	rpcOpts := append(c.copyRPCOpts(), rpc.WithStaticAuthenticationMaterial(c.conf.Auth.AccessToken))
+
+	conn, err := rpc.DialDirectGRPC(
+		c.c.Context,
+		c.baseURL.Host,
+		nil,
+		rpcOpts...,
+	)
+	if err != nil {
+		return err
+	}
+
+	c.client = apppb.NewAppServiceClient(conn)
+	c.dataClient = datapb.NewDataServiceClient(conn)
+	return nil
+}
+
+// logout logs out the client and clears the config.
+func (c *appClient) logout() error {
+	if err := removeConfigFromCache(); err != nil && !os.IsNotExist(err) {
+		return err
+	}
+	c.conf = &config{}
+	return nil
+}
+
+func (c *appClient) prepareDial(
+	orgStr, locStr, robotStr, partStr string,
+	debug bool,
+) (context.Context, string, []rpc.DialOption, error) {
+	if err := c.ensureLoggedIn(); err != nil {
+		return nil, "", nil, err
+	}
+	if err := c.selectOrganization(orgStr); err != nil {
+		return nil, "", nil, err
+	}
+	if err := c.selectLocation(locStr); err != nil {
+		return nil, "", nil, err
+	}
+
+	part, err := c.robotPart(c.selectedOrg.Id, c.selectedLoc.Id, robotStr, partStr)
+	if err != nil {
+		return nil, "", nil, err
+	}
+
+	rpcDialer := rpc.NewCachedDialer()
+	defer func() {
+		utils.UncheckedError(rpcDialer.Close())
+	}()
+	dialCtx := rpc.ContextWithDialer(c.c.Context, rpcDialer)
+
+	rpcOpts := append(c.copyRPCOpts(),
+		rpc.WithExternalAuth(c.baseURL.Host, part.Fqdn),
+		rpc.WithStaticExternalAuthenticationMaterial(c.conf.Auth.AccessToken),
+	)
+
+	if debug {
+		rpcOpts = append(rpcOpts, rpc.WithDialDebug())
+	}
+
+	return dialCtx, part.Fqdn, rpcOpts, nil
 }
 
 func (t *token) isExpired() bool {
